# app/routers/parking_lots.py
from datetime import datetime

from fastapi import APIRouter, Depends, Query, status

from app.deps import get_current_admin_user
from app.models.admin_user import AdminUser
from app.schemas.event import EventOut
from app.schemas.parking_lot import ParkingLotCreate, ParkingLotOut, ParkingLotUpdate
from app.usecases.parking_lot_usecase import ParkingLotUsecase, get_parking_lot_usecase

router = APIRouter(prefix="/parking-lots", tags=["parking-lots"])


@router.post("", response_model=ParkingLotOut, status_code=status.HTTP_201_CREATED, summary="駐車場登録")
def create_parking_lot(
    payload: ParkingLotCreate,
    usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase),
    _admin: AdminUser = Depends(get_current_admin_user),
):
    """新しい駐車場を登録する。"""
    return usecase.create_parking_lot(name=payload.name, capacity=payload.capacity)


@router.get("", response_model=list[ParkingLotOut], summary="駐車場一覧取得")
def list_parking_lots(usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase)):
    """登録済みの駐車場と、それぞれの現在の駐車台数を一覧で返す。"""
    return usecase.list_parking_lots()


@router.get("/{lot_id}", response_model=ParkingLotOut, summary="駐車場詳細取得")
def get_parking_lot(lot_id: int, usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase)):
    """指定した駐車場の詳細（現在の駐車台数を含む）を返す。"""
    return usecase.get_parking_lot(lot_id)


@router.patch("/{lot_id}", response_model=ParkingLotOut, summary="駐車場情報の更新")
def update_parking_lot(
    lot_id: int,
    payload: ParkingLotUpdate,
    usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase),
    _admin: AdminUser = Depends(get_current_admin_user),
):
    """駐車場名・収容台数を更新する。指定したフィールドのみ更新される。"""
    return usecase.update_parking_lot(lot_id, name=payload.name, capacity=payload.capacity)


@router.delete("/{lot_id}", status_code=status.HTTP_204_NO_CONTENT, summary="駐車場削除")
def delete_parking_lot(
    lot_id: int,
    usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase),
    _admin: AdminUser = Depends(get_current_admin_user),
):
    """駐車場を削除する。紐づくデバイスが存在する場合は409エラーになる（先にデバイスを削除・移動すること）。"""
    usecase.delete_parking_lot(lot_id)


@router.get("/{lot_id}/events", response_model=list[EventOut], summary="駐車場の入出庫履歴取得")
def list_parking_lot_events(
    lot_id: int,
    since: datetime | None = Query(default=None, description="この日時以降に検出されたイベントに絞り込む"),
    until: datetime | None = Query(default=None, description="この日時以前に検出されたイベントに絞り込む"),
    limit: int = Query(default=100, le=1000, description="取得件数の上限（最大1000）"),
    usecase: ParkingLotUsecase = Depends(get_parking_lot_usecase),
    _admin: AdminUser = Depends(get_current_admin_user),
):
    """指定した駐車場に紐づく入出庫イベントを検出日時の新しい順に返す。"""
    return usecase.list_events(lot_id, since=since, until=until, limit=limit)