# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trapa Center API",
    "description": "駐車場入出庫トラッキングシステム「トラパ」のセンターAPI。エッジデバイスからの入出庫イベント登録、ヘルスチェック、デバイスへのコマンド（再起動など）配信をポーリング方式で提供する。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/parking-lots": {
      "get": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場一覧取得",
        "description": "登録済みの駐車場と、それぞれの現在の駐車台数を一覧で返す。",
        "operationId": "list_parking_lots_api_v1_parking_lots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParkingLotOut"
                  },
                  "type": "array",
                  "title": "Response List Parking Lots Api V1 Parking Lots Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場登録",
        "description": "新しい駐車場を登録する。",
        "operationId": "create_parking_lot_api_v1_parking_lots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParkingLotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkingLotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/parking-lots/{lot_id}": {
      "get": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場詳細取得",
        "description": "指定した駐車場の詳細（現在の駐車台数を含む）を返す。",
        "operationId": "get_parking_lot_api_v1_parking_lots__lot_id__get",
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkingLotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場情報の更新",
        "description": "駐車場名・収容台数を更新する。指定したフィールドのみ更新される。",
        "operationId": "update_parking_lot_api_v1_parking_lots__lot_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParkingLotUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkingLotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場削除",
        "description": "駐車場を削除する。紐づくデバイスが存在する場合は409エラーになる（先にデバイスを削除・移動すること）。",
        "operationId": "delete_parking_lot_api_v1_parking_lots__lot_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parking-lots/{lot_id}/events": {
      "get": {
        "tags": [
          "parking-lots"
        ],
        "summary": "駐車場の入出庫履歴取得",
        "description": "指定した駐車場に紐づく入出庫イベントを検出日時の新しい順に返す。",
        "operationId": "list_parking_lot_events_api_v1_parking_lots__lot_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lot Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "この日時以降に検出されたイベントに絞り込む",
              "title": "Since"
            },
            "description": "この日時以降に検出されたイベントに絞り込む"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "この日時以前に検出されたイベントに絞り込む",
              "title": "Until"
            },
            "description": "この日時以前に検出されたイベントに絞り込む"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "description": "取得件数の上限（最大1000）",
              "default": 100,
              "title": "Limit"
            },
            "description": "取得件数の上限（最大1000）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Parking Lot Events Api V1 Parking Lots  Lot Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_id": {
            "type": "integer",
            "title": "Device Id",
            "description": "このイベントを登録したデバイスのID"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "イベント種別（entry=入庫 / exit=出庫）"
          },
          "vehicle_track_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Track Id"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At",
            "description": "エッジデバイスが検出した日時"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At",
            "description": "サーバーがイベントを受信した日時"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "event_type",
          "vehicle_track_id",
          "detected_at",
          "received_at"
        ],
        "title": "EventOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParkingLotCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "駐車場名"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity",
            "description": "収容台数"
          }
        },
        "type": "object",
        "required": [
          "name",
          "capacity"
        ],
        "title": "ParkingLotCreate"
      },
      "ParkingLotOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "current_count": {
            "type": "integer",
            "title": "Current Count",
            "description": "現在の駐車台数。入出庫イベントごとに増減する"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "capacity",
          "current_count",
          "created_at"
        ],
        "title": "ParkingLotOut"
      },
      "ParkingLotUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "駐車場名"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity",
            "description": "収容台数"
          }
        },
        "type": "object",
        "title": "ParkingLotUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin console access token (from /auth/login)",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Edge device API key",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}